# src/pkg/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The IPEHR Gateway is an openEHR compliant EHR server implementation that stores encrypted medical data in a Filecoin distributed file storage.",
        "title": "IPEHR Gateway API",
        "contact": {
            "name": "API Support",
            "url": "https://bsn.si/blockchain",
            "email": "support@bsn.si"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.2"
    },
    "host": "gateway.ipehr.org",
    "basePath": "/v1",
    "paths": {
...
        "/ehr": {
            "get": {
                "description": "Retrieve the EHR with the specified subject_id and subject_namespace.\nThese subject parameters will be matched against EHR’s\nEHR_STATUS.subject.external_ref.id.value and EHR_STATUS.subject.external_ref.namespace values.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EHR"
                ],
                "summary": "Get EHR summary by subject id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "subject id. Example: ins01",
                        "name": "subject_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "id namespace. Example: examples",
                        "name": "subject_namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer AccessToken",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UserId",
                        "name": "AuthUserId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The identifier of the system, typically a reverse domain identifier",
                        "name": "EhrSystemId",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.EhrSummary"
                        }
                    },
                    "202": {
                        "description": "Is returned when the request is still being processed"
                    },
                    "400": {
                        "description": "Is returned when userID is empty"
                    },
                    "404": {
                        "description": "Is returned when an EHR with ehr_id does not exist."
                    }
                }
            },
            "post": {
                "description": "Create a new EHR with an auto-generated identifier.\nAn EHR_STATUS resource needs to be always created and committed in the new EHR. This resource MAY be also supplied by the client as the request body. If not supplied, a default EHR_STATUS will be used by the service with following attributes:\n- `is_queryable`: true\n- `is_modifiable`: true\n- `subject`: a PARTY_SELF object\n\nAll other required EHR attributes and resources will be automatically created as needed by the [EHR creation semantics](https://specifications.openehr.org/releases/RM/latest/ehr.html#_ehr_creation_semantics).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EHR"
                ],
                "summary": "Create EHR",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer AccessToken",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UserId",
                        "name": "AuthUserId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The identifier of the system, typically a reverse domain identifier",
                        "name": "EhrSystemId",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "GroupAccessId - UUID. If not specified, the default access group will be used.",
                        "name": "GroupAccessId",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "The new EHR resource is returned in the body when the request’s `Prefer` header value is `return=representation`, otherwise only headers are returned.",
                        "name": "Prefer",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Query Request",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.EhrCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.EhrSummary"
                        },
                        "headers": {
                            "ETag": {
                                "type": "string",
                                "description": "ehr_id of created document. Example: 7d44b88c-4199-4bad-97dc-d78268e01398"
                            },
                            "Location": {
                                "type": "string",
                                "description": "{baseUrl}/ehr/7d44b88c-4199-4bad-97dc-d78268e01398"
                            },
                            "RequestID": {
                                "type": "string",
                                "description": "Request identifier"
                            }
                        }
                    },
                    "400": {
                        "description": "Is returned when the request body (if provided)  could      not  be  parsed."
                    },
                    "409": {
                        "description": "Unable to create a new EHR due to a conflict with an already existing EHR with the same subject id, namespace pair."
                    },
                    "500": {
                        "description": "Is returned when an unexpected error occurs while processing a request"
                    }
                }
            }
        },
...
        "/ehr/{ehr_id}/ehr_status": {
            "get": {
                "description": "Retrieves a version of the EHR_STATUS associated with the EHR identified by `ehr_id`. If `version_at_time` is supplied, retrieves the version extant at specified time, otherwise retrieves the latest EHR_STATUS version.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EHR_STATUS"
                ],
                "summary": "Get EHR_STATUS version by time",
                "parameters": [
                    {
                        "type": "string",
                        "description": "EHR identifier taken from EHR.ehr_id.value. Example: 7d44b88c-4199-4bad-97dc-d78268e01398",
                        "name": "ehr_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "A given time in the extended ISO 8601 format. Example: 2015-01-20T19:30:22.765+01:00",
                        "name": "version_at_time",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer AccessToken",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UserId",
                        "name": "AuthUserId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The identifier of the system, typically a reverse domain identifier",
                        "name": "EhrSystemId",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.EhrStatusUpdate"
                        }
                    },
                    "202": {
                        "description": "Is returned when the request is still being processed"
                    },
                    "400": {
                        "description": "Is returned when the request has invalid content such as an invalid `version_at_time` format."
                    },
                    "404": {
                        "description": "Is returned when EHR with `ehr_id` does not exist or a version of an EHR_STATUS resource does not exist at the specified `version_at_time`"
                    },
                    "500": {
                        "description": "Is returned when an unexpected error occurs while processing a request"
                    }
                }
            },
            "put": {
                "description": "Updates EHR_STATUS associated with the EHR identified by `ehr_id`.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EHR_STATUS"
                ],
                "summary": "Update EHR_STATUS",
                "parameters": [
                    {
                        "type": "string",
                        "description": "EHR identifier. Example: 7d44b88c-4199-4bad-97dc-d78268e01398",
                        "name": "ehr_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer AccessToken",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UserId",
                        "name": "AuthUserId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The identifier of the system, typically a reverse domain identifier",
                        "name": "EhrSystemId",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "The existing latest `version_uid` of EHR_STATUS resource (i.e. the `preceding_version_uid`)  must  be  specified.",
                        "name": "If-Match",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Updated resource is returned in the body when the request’s `Prefer` header value is `return=representation`, otherwise only headers are returned.",
                        "name": "Prefer",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "EHR_STATUS",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.EhrStatusUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.EhrStatusUpdate"
                        },
                        "headers": {
                            "ETag": {
                                "type": "string",
                                "description": "uid of created document. Example: 8849182c-82ad-4088-a07f-48ead4180515::openEHRSys.example.com::2"
                            },
                            "Location": {
                                "type": "string",
                                "description": "{baseUrl}/ehr/7d44b88c-4199-4bad-97dc-d78268e01398/ehr_status/8849182c-82ad-4088-a07f-48ead4180515::openEHRSys.example.com::2"
                            },
                            "RequestID": {
                                "type": "string",
                                "description": "Request identifier"
                            }
                        }
                    },
                    "204": {
                        "description": "Is returned when `Prefer` header is missing or is set to `return=minimal`",
                        "headers": {
                            "ETag": {
                                "type": "string",
                                "description": "uid of created document. Example: 8849182c-82ad-4088-a07f-48ead4180515::openEHRSys.example.com::2"
                            },
                            "Location": {
                                "type": "string",
                                "description": "{baseUrl}/ehr/7d44b88c-4199-4bad-97dc-d78268e01398/ehr_status/8849182c-82ad-4088-a07f-48ead4180515::openEHRSys.example.com::2"
                            }
                        }
                    },
                    "400": {
                        "description": "Is returned when the request has invalid content."
                    },
                    "404": {
                        "description": "Is returned when an EHR with ehr_id does not exist."
                    },
                    "412": {
                        "description": "Is returned when `If-Match` request header doesn’t match the latest version on the service side. Returns also latest `version_uid` in the `Location` and `ETag` headers."
                    },
                    "500": {
                        "description": "Is returned when an unexpected error occurs while processing a request"
                    }
                }
            }
        },
        "/ehr/{ehr_id}/ehr_status/{version_uid}": {
            "get": {
                "description": "Retrieves a particular version of the EHR_STATUS identified by `version_uid` and associated with the EHR identified by `ehr_id`.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EHR_STATUS"
                ],
                "summary": "Get EHR_STATUS by version id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "EHR identifier taken from EHR.ehr_id.value. Example: 7d44b88c-4199-4bad-97dc-d78268e01398",
                        "name": "ehr_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "VERSION identifier taken from VERSION.uid.value. Example: 8849182c-82ad-4088-a07f-48ead4180515::openEHRSys.example.com::2",
                        "name": "version_uid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer AccessToken",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UserId",
                        "name": "AuthUserId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The identifier of the system, typically a reverse domain identifier",
                        "name": "EhrSystemId",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.EhrStatusUpdate"
                        }
                    },
                    "202": {
                        "description": "Is returned when the request is still being processed"
                    },
                    "400": {
                        "description": "Is returned when AuthUserId is not specified"
                    },
                    "404": {
                        "description": "is returned when an EHR with `ehr_id` does not exist or when an EHR_STATUS with `version_uid` does not exist."
                    },
                    "500": {
                        "description": "Is returned when an unexpected error occurs while processing a request"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.EhrCreateRequest": {
            "type": "object",
            "properties": {
                "_type": {
                    "type": "string"
                },
                "archetype_node_id": {
                    "type": "string"
                },
                "isModifiable": {
                    "type": "boolean"
                },
                "isQueryable": {
                    "type": "boolean"
                },
                "name": {
                    "type": "object",
                    "properties": {
                        "value": {
                            "type": "string"
                        }
                    }
                },
                "subject": {
                    "type": "object",
                    "properties": {
                        "external_ref": {
                            "$ref": "#/definitions/model.ExternalRef"
                        }
                    }
                }
            }
        },
...
        "model.EhrStatusUpdate": {
            "type": "object",
            "properties": {
                "_type": {
                    "type": "string"
                },
                "archetype_node_id": {
                    "type": "string"
                },
                "is_modifiable": {
                    "type": "boolean"
                },
                "is_queryable": {
                    "type": "boolean"
                },
                "name": {
                    "type": "object",
                    "properties": {
                        "value": {
                            "type": "string"
                        }
                    }
                },
                "other_details": {
                    "type": "object",
                    "properties": {
                        "_type": {
                            "type": "string"
                        },
                        "archetype_node_id": {
                            "type": "string"
                        },
                        "items": {
                            "type": "array",
                            "items": {}
                        },
                        "name": {
                            "type": "object",
                            "properties": {
                                "value": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "subject": {
                    "type": "object",
                    "properties": {
                        "external_ref": {
                            "$ref": "#/definitions/model.ExternalRef"
                        }
                    }
                },
                "uid": {
                    "type": "object",
                    "properties": {
                        "_type": {
                            "type": "string"
                        },
                        "value": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "model.EhrSummary": {
            "type": "object",
            "properties": {
                "ehr_access": {
                    "type": "object",
                    "properties": {
                        "id": {
                            "type": "string"
                        },
                        "namespace": {
                            "type": "string"
                        },
                        "type": {
                            "type": "string"
                        }
                    }
                },
                "ehr_id": {
                    "type": "string"
                },
                "ehr_status": {
                    "type": "object",
                    "properties": {
                        "id": {
                            "type": "string"
                        },
                        "namespace": {
                            "type": "string"
                        },
                        "type": {
                            "type": "string"
                        }
                    }
                },
                "system_id": {
                    "type": "string"
                },
                "time_created": {
                    "type": "object",
                    "properties": {
                        "value": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "model.ExternalRef": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "object",
                    "properties": {
                        "_type": {
                            "type": "string"
                        },
                        "scheme": {
                            "type": "string"
                        },
                        "value": {
                            "type": "string"
                        }
                    }
                },
                "namespace": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
    }
}